Move wrench package to internal/wrench
