+++ cmd/jujud/agent/machine.go
@@ -73,6 +73,7 @@ import (
 	"github.com/juju/juju/internal/pubsub/centralhub"
 	"github.com/juju/juju/internal/service"
 	"github.com/juju/juju/internal/storage/looputil"
+	"github.com/juju/juju/internal/wrench"
 	jujunames "github.com/juju/juju/juju/names"
 	"github.com/juju/juju/rpc/params"
 	"github.com/juju/juju/state"
@@ -90,7 +91,6 @@ import (
 	psworker "github.com/juju/juju/worker/pubsub"
 	"github.com/juju/juju/worker/servicefactory"
 	"github.com/juju/juju/worker/upgradesteps"
-	"github.com/juju/juju/wrench"
 )
 
 var (
