+++ internal/wrench/wrench_nix_test.go
@@ -10,7 +10,7 @@ import (
 
 	gc "gopkg.in/check.v1"
 
-	"github.com/juju/juju/wrench"
+	"github.com/juju/juju/internal/wrench"
 )
 
 const fileNotFound = `stat .+: no such file or directory`
