+++ internal/wrench/wrench_test.go
@@ -13,8 +13,8 @@ import (
 	jc "github.com/juju/testing/checkers"
 	gc "gopkg.in/check.v1"
 
+	"github.com/juju/juju/internal/wrench"
 	coretesting "github.com/juju/juju/testing"
-	"github.com/juju/juju/wrench"
 )
 
 func TestPackage(t *stdtesting.T) {
