+++ worker/instancemutater/mutater.go
@@ -22,8 +22,8 @@ import (
 	"github.com/juju/juju/core/status"
 	"github.com/juju/juju/core/watcher"
 	"github.com/juju/juju/environs"
+	"github.com/juju/juju/internal/wrench"
 	"github.com/juju/juju/rpc/params"
-	"github.com/juju/juju/wrench"
 )
 
 //go:generate go run go.uber.org/mock/mockgen -package mocks -destination mocks/mutatercontext_mock.go github.com/juju/juju/worker/instancemutater MutaterContext
