+++ worker/migrationmaster/worker.go
@@ -26,9 +26,9 @@ import (
 	"github.com/juju/juju/core/watcher"
 	"github.com/juju/juju/internal/migration"
 	"github.com/juju/juju/internal/tools"
+	"github.com/juju/juju/internal/wrench"
 	"github.com/juju/juju/rpc/params"
 	"github.com/juju/juju/worker/fortress"
-	"github.com/juju/juju/wrench"
 )
 
 var (
