+++ worker/storageprovisioner/volume_events.go
@@ -12,8 +12,8 @@ import (
 	"github.com/juju/juju/core/watcher"
 	"github.com/juju/juju/internal/storage"
 	"github.com/juju/juju/internal/storage/plans"
+	"github.com/juju/juju/internal/wrench"
 	"github.com/juju/juju/rpc/params"
-	"github.com/juju/juju/wrench"
 )
 
 // volumesChanged is called when the lifecycle states of the volumes
